Revert "move `hamMenu` in `pages`"
